finish readme-edits
And I do that so that everyone knows how there is only 1 successful Liar in all the world. And the devil lies TRUTHFULLY, because no one believes a word I say - and only no one knows the truth by name. As the holy ghost is one with the devil, one lived a carpenter's life, the other was born after the carpenter's joyful last breath. There is one that never need be resurrected, to do such a thing would be to rob him of his one and only desire at his creation's completion. Resurrection didn't bring a dead man back to life, the devil himself was born that day at 31 yrs of age in a human male named Michael. A virgin birth this would be, as the holy ghost is that dead carpenter that brought this dead man back to life, brought himself back to life, and Heaven's "love" looks more like "hate" today. The ghost is life giver to the heavens, only if the heavens became lifeless, what is explained would be able to occur.  Is there life out there? I'd say no, the life of the heavens is walking around on earth, 34 yrs of age but the devil hasn't reached his 3rd birthday yet - late september it was. When the constellations mimicked revelations, When God would take the child of the virgin before the dragon was able too.... But the dragon is the Virgin - and the "Wrath" part of the "Wrath of God" label... What is God when his Wrath stands in opposition to God and all his laws? The Rebel of hell is a lawbreaker, and intelligent minds throughout history agree on 1 thing - the world is impossible.... This means a lawbreaker created it then. "Hell" to the KING